src/detection_model_service/src/requirements.txt
python-dotenv==1.0.1
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchvision==0.22.0
tqdm==4.67.1
ultralytics==8.3.127
ultralytics-thop==2.0.14
numpy==2.2.5
confluent-kafka==2.8.2
opencv-python
pillow==11.2.1